trying

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -6,7 +6,7 @@
 
 int main(void)
 {
-	long prime = 612852475143, div;
+	int prime = 612852475143, div;
 
 	while (div < (prime / 2))
 	{
@@ -15,7 +15,7 @@ int main(void)
 			prime /= 2;
 			continue;
 		}
-		for (div = 3; div < (prime / 2); div+= 2)
+		for (div = 3; div < (prime / 2); div += 2)
 		{
 			if (prime % div == 0)
 			{
